Store ciphertext size in the DB instead of the plaintext size
This is useful as a first cheap check to see if backup data on the backend still has the expected size.
The plaintext size wasn't really used for anything anyway. Existing installation that still store the plaintext size will log different file size as suspicious and prefer those chunks in checks, but there shouldn't be other negative consequences from this.